Refactor: 메인 페이지 컴포넌트 파일 이동 #10
